+++ src/consts/exceptions.conts.ts
@@ -28,5 +28,11 @@ export const NRG_EXCEPTIONS = {
         message: () => {
             return `The requested item-type does not exist in the default types, nor in the additional types.`;
         }
+    },
+    NoValidFileTemplateForThisItemException: {
+        name: 'NoValidFileTemplateForThisItemException',
+        message: (itemName: string) => {
+            return `You specified in the configuration that a custom file template should be used for "${itemName}" items, but neither the template field or the templateFile leads to a valid file, or the file is empty.`;
+        }
     }
 }