+++ src/interfaces/conf-reader.interface.ts
@@ -1,8 +1,12 @@
-import { IEnergyAdditionalType } from './energy-cli-conf.interface';
+
+import { IAdditionalType } from './additional-type.interface';
+import { ICustomFileTemplate } from './custom-file-template.interface';
 
 export interface IConfReader {
     getSrcFolder(): string;
     getDefaultFilesExt(): string;
+    getAdditionalTypes(): IAdditionalType[];
+    getCustomFileTemplates(): ICustomFileTemplate[];
     useDotnetInterfaceStyle(): boolean;
-    getAdditionalTypes(): IEnergyAdditionalType[];
+    getDefaultProjectStructure(): any;
 }