+++ src/interfaces/custom-file-template.interface.ts
@@ -1,0 +1,5 @@
+export interface ICustomFileTemplate {
+    itemName: string;
+    template?: string;
+    templateFile?: string;
+}